chore(core): merge fallout

+++ core/src/main/java/io/questdb/cairo/CairoConfiguration.java
@@ -229,15 +229,15 @@ public interface CairoConfiguration {
 
     int getSqlJitBindVarsMemoryMaxPages();
 
-    long getSqlJitBindVarsMemoryPageSize();
+    int getSqlJitBindVarsMemoryPageSize();
 
     int getSqlJitIRMemoryMaxPages();
 
-    long getSqlJitIRMemoryPageSize();
+    int getSqlJitIRMemoryPageSize();
 
     int getSqlJitMode();
 
-    long getSqlJitPageAddressCacheThreshold();
+    int getSqlJitPageAddressCacheThreshold();
 
     int getSqlJitRowsThreshold();
